add check for .editable when deleting npmessage

+++ events/player_events/trackEnd.js
@@ -1,7 +1,7 @@
 const Event = require("../../structures/event.js");
 
 module.exports = new Event("trackEnd", async (player, queue, track) => {
-    if(queue.npmessage) {
+    if(queue.npmessage && queue.npmessage.editable) {
         queue.npmessage.delete().catch(error=> {});
     }
